refactor: add dialog container
