+++ src/components/Dialog/index.ts
@@ -1,0 +1,3 @@
+export * from './Message';
+export * from './Settings';
+export * from './Container';
