+++ src/components/index.ts
@@ -5,4 +5,3 @@ export * from './Footer';
 export * from './Layout';
 export * from './Success';
 export * from './UserForm';
-export * from './MessageModal';
