+++ src/hooks/useUser.ts
@@ -1,0 +1,10 @@
+import { useQuery } from 'react-query';
+import { getUser } from '@/api';
+
+export const useUser = (username: string) => {
+  const { data } = useQuery(['user', { username }], () =>
+    getUser({ username })
+  );
+
+  return data?.user;
+};
